finish DiaryController interface test case.

+++ server/source/mmdiary/src/main/java/com/sdsxer/mmdiary/web/DiaryController.java
@@ -225,7 +225,7 @@ public class DiaryController extends BaseController {
     }
 
     // delete old cover image
-    if(coverImageUpdated) {
+    if(coverImageUpdated && !StringUtils.isEmpty(originalCoverImagePath)) {
       try {
         Files.delete(Paths.get(storageService.getRootLocation()
             + File.separator + originalCoverImagePath));
